float

+++ 離散數學/遞迴式(硬算).py
@@ -10,16 +10,16 @@ def fff(x, a1,a2,a3, p1,p2, ex1,exp, f0, f1, f2):
 
 print("f(x) = a1*f(x-1) + a2*f(x-2) + a3*f(x-3) + p1*x + p2 + ex1*exp^x")
 num_x = int(input("第x項 = "))
-num_a1 = int(input("係數a1 = "))
-num_a2 = int(input("係數a2 = "))
-num_a3 = int(input("係數a3 = "))
-num_p1 = int(input("係數p1 = "))
-num_p2 = int(input("係數p2 = "))
-num_ex1 = int(input("係數ex1 = "))
-num_exp = int(input("指數底exp = "))
-num_f0 = int(input("初始值f(0) = "))
-num_f1 = int(input("初始值f(1) = "))
-num_f2 = int(input("初始值f(2) = "))
+num_a1 = float(input("係數a1 = "))
+num_a2 = float(input("係數a2 = "))
+num_a3 = float(input("係數a3 = "))
+num_p1 = float(input("係數p1 = "))
+num_p2 = float(input("係數p2 = "))
+num_ex1 = float(input("係數ex1 = "))
+num_exp = float(input("指數底exp = "))
+num_f0 = float(input("初始值f(0) = "))
+num_f1 = float(input("初始值f(1) = "))
+num_f2 = float(input("初始值f(2) = "))
 
 
 
@@ -40,15 +40,15 @@ def fff(x, a1,a2, ex1,exp, f0, f1):
     return result
 
 print("f(x) = a1*f(x-1) + a2*f(x-2) + ex1*exp^x")
-num_x = int(input("第x項 = "))
-num_a1 = int(input("係數a1 = "))
-num_a2 = int(input("係數a2 = "))
+num_x = float(input("第x項 = "))
+num_a1 = float(input("係數a1 = "))
+num_a2 = float(input("係數a2 = "))
 
-num_ex1 = int(input("係數ex1 = "))
-num_exp = int(input("指數底exp = "))
+num_ex1 = float(input("係數ex1 = "))
+num_exp = float(input("指數底exp = "))
 
-num_f0 = int(input("初始值f(0) = "))
-num_f1 = int(input("初始值f(1) = "))
+num_f0 = float(input("初始值f(0) = "))
+num_f1 = float(input("初始值f(1) = "))
 
 
 fn = fff(num_x, num_a1, num_a2, num_ex1, num_exp, num_f0, num_f1)
